server: move the main tenant orchestration logic to its own file
This moves `(o *channelOrchestrator) startControlledServer` alongside
the other methods of `channelOrchestrator`.
There is further no code change.

Release note: None
